Create 15_craywatch_maps.R
Remove manual addition of point, if this point is registered through waarnemingen, it is considered a waarnemingen punt, and not a craywatch point. Changing this logic is not advisable. If you still want to add it as craywatch, you'll have to fill it in the CW-analysis dataset, and this will also add up in numbers to other scripts. (and will be filtered out based on waarnemingenID from the GBIF).